requirements.txt
numpy==1.26.4
pandas==2.1.4
pdfplumber==0.10.3
tabula-py==2.9.0
dbfread==2.0.7
openpyxl==3.1.2
